For #975: Uses failable constructor
